requirements.txt
numpy
pandas
langchain
langchain_openai
langchain_mistralai
langchain_anthropic
mistralai